Controller und Datenbanken Prügelei

+++ stepflowBackend/src/main/java/de/sommer/stepflowBackend/controller/UserController.java
@@ -1,15 +1,25 @@
 package de.sommer.stepflowBackend.controller;
 
-import de.sommer.stepflowBackend.models.User;
-import de.sommer.stepflowBackend.services.api.UserService;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-import java.util.Optional;
+import de.sommer.stepflowBackend.dto.UserDTO;
+import de.sommer.stepflowBackend.dto.UserListDTO;
+import de.sommer.stepflowBackend.models.User;
+import de.sommer.stepflowBackend.services.api.AuthService;
+import de.sommer.stepflowBackend.services.api.UserService;
 
 @RestController
 @RequestMapping("/api/users")
@@ -21,25 +31,43 @@ public class UserController {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
+    @Autowired 
+    private AuthService authService;
+
     @GetMapping
-    public List<User> getAllUsers() {
-        return userService.getAllUsers();
+    public ResponseEntity<?> getAllUsers() {
+        UserListDTO userList = UserListDTO.onlyName(userService.getAllUsers());
+        return ResponseEntity.ok(userList);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable Long id) {
+    public ResponseEntity<?> getUserById(@PathVariable int id, @RequestHeader("Authorization") String token) {
+        Optional<User> currentUser = userService.getUserByEmail(authService.extractEmail(token.replace("Bearer ", "")));
         Optional<User> user = userService.getUserById(id);
-        return user.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        if (user.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        if (currentUser.get().getId() != id && !currentUser.get().getRole().equals("admin")) {
+            return ResponseEntity.ok(UserDTO.onlyName(user.get()));
+        }
+        else {
+            return ResponseEntity.ok(new UserDTO(user.get()));
+        }
+        
     }
 
     @PostMapping
-    public User createUser(@RequestBody User user) {
+    public ResponseEntity<?> createUser(@RequestBody UserDTO user) {
         user.setPassword(passwordEncoder.encode(user.getPassword()));
-        return userService.createUser(user);
+        return ResponseEntity.ok(new UserDTO(userService.createUser(new User(user))));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails) {
+    public ResponseEntity<?> updateUser(@PathVariable int id, @RequestBody User userDetails, @RequestHeader("Authorization") String token) {
+        Optional<User> user = userService.getUserByEmail(authService.extractEmail(token.replace("Bearer ", "")));
+        if (user.isEmpty() || user.get().getId() != id && !user.get().getRole().equals("admin")) {
+            return ResponseEntity.status(403).build();
+        }
         userDetails.setPassword(passwordEncoder.encode(userDetails.getPassword()));
         try {
             User updatedUser = userService.updateUser(id, userDetails);
@@ -50,7 +78,11 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<?> deleteUser(@PathVariable int id, @RequestHeader("Authorization") String token) {
+        Optional<User> user = userService.getUserByEmail(authService.extractEmail(token.replace("Bearer ", "")));
+        if (user.isEmpty() || user.get().getId() != id && !user.get().getRole().equals("admin")) {
+            return ResponseEntity.status(403).build();
+        }
         userService.deleteUser(id);
         return ResponseEntity.noContent().build();
     }
